FROMLIST: arm64: dts: meson-sm1: specify suspend OPP
Tag the 1,2GHz OPP as suspend OPP to be set before going in suspend mode.

It has been reported that using various OPPs can lead to error or
resume with a different OPP from the ROM, thus use this safe OPP as
it is the default OPP used by the BL2 boot firmware.

Signed-off-by: Neil Armstrong <narmstrong@baylibre.com>
Link: https://lore.kernel.org/linux-amlogic/20190827100307.21661-3-narmstrong@baylibre.com/
Change-Id: I7a039fb45837402d6821d3867112ea20cd4d963e